backend/requirements.txt
django==3.1.7
djangorestframework==3.12.4
django-allauth==0.44.0
dj-rest-auth==2.1.4
django-cors-headers==3.7.0
djangorestframework-camel-case==1.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
sentry-sdk==1.0.0
requests==2.31.0
python-dotenv==0.19.0
tinkoff-invest==1.0.5
pandas>=1.3.0
backtrader==1.9.78.123
numpy==1.24.3
matplotlib==3.7.3
seaborn==0.12.2
ta==0.10.2
TA-lib==0.6.3
tinkoff-investments==0.2.0b111
pytz>=2023.3
pytest==8.3.5